fix #9243 feat(cirrus): Node 18, proxy and remove extra files
