+++ demo-app/frontend/Dockerfile
@@ -1,4 +1,4 @@
-FROM node:14 as build
+FROM node:18 as build
 
 WORKDIR /app
 
@@ -10,9 +10,4 @@ COPY . .
 
 RUN npm run build
 
-FROM nginx:alpine
-
-COPY --from=build /app/build /usr/share/nginx/html
-
-EXPOSE 80
-CMD ["nginx", "-g", "daemon off;"]
+CMD ["npm", "start"]
