--- demo-app/frontend/src/App.js
@@ -1,21 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import './App.css';
-
-function App() {
-  const [message, setMessage] = useState('');
-
-  useEffect(() => {
-    fetch('http://localhost:3002/api/data')
-      .then(response => response.json())
-      .then(data => setMessage(data.message))
-      .catch(error => console.error('Error fetching data:', error));
-  }, []);
-
-  return (
-    <div className="App">
-      <h1>{message}</h1>
-    </div>
-  );
-}
-
-export default App;
