+++ demo-app/frontend/src/setupProxy.js
@@ -1,0 +1,12 @@
+// src/setupProxy.js
+const { createProxyMiddleware } = require('http-proxy-middleware');
+
+module.exports = function(app) {
+  app.use(
+    '/api',
+      createProxyMiddleware({
+      target: process.env.DEMO_APP_SERVER??'http://localhost:3002',
+      changeOrigin: true,
+    })
+  );
+};
